Return an exception message to stderr
Fully-formatted exception as CSS with
a backtrace is useless to have useful
comparison against known errors we
might want to test against.

This is also inconsistent with regards
to other engines invoked via the command-line
(sassc, Ruby sass CLI) - since in that
case the exception is printed on
the standard output.

Therefore return empty string on
the standard output and only an
error message (without backtrace) to stderr.

Fixes: #497 (to some extent)